used JOINS

+++ Controller/UserController.js
@@ -1,18 +1,18 @@
 const express = require("express");
 const router = express.Router();
 const mysqlConnection = require("../Config/cnctDb");
-var md5 = require('md5');
+var md5 = require("md5");
 
 //Get all User
 exports.GetAll = (req, res) => {
-  mysqlConnection.query("SELECT * FROM users", (err, rows, fields) => {
+  var qury = "SELECT users.id,name,email,salary FROM users LEFT JOIN payroll ON(users.id = payroll.userId)";
+  mysqlConnection.query(qury, (err, rows, fields) => {
     if (!err) {
       res.status(200).json({
-        success:true,
-        data: rows
-      })
-    }
-    else{
+        success: true,
+        data: rows,
+      });
+    } else {
       res.status(200).json({
         success: false,
         data: { message: "Something went to wrong!" + err },
@@ -24,20 +24,19 @@ exports.GetAll = (req, res) => {
 //Get an User By ID
 exports.GetById = (req, res) => {
   mysqlConnection.query(
-    "SELECT * FROM users WHERE id = ?",
+    "SELECT * FROM users WHERE id = ? LEFT JOIN payroll ON(users.id = payroll.userId)",
     [req.params.id],
     (err, rows, fields) => {
-      if (!err){
-         res.status(200).json({
-        success:true,
-        data: rows
-      })
-      }
-      else{
+      if (!err) {
         res.status(200).json({
-        success: false,
-        data: { message: "Something went to wrong!" + err },
-      });
+          success: true,
+          data: rows,
+        });
+      } else {
+        res.status(200).json({
+          success: false,
+          data: { message: "Something went to wrong!" + err },
+        });
       }
     }
   );
@@ -46,20 +45,19 @@ exports.GetById = (req, res) => {
 //Delete an User
 exports.DeleteById = (req, res) => {
   mysqlConnection.query(
-    "DELETE FROM users WHERE id = ?",
+    "DELETE FROM users WHERE id = ? LEFT JOIN payroll ON(users.id = payroll.userId)",
     [req.params.id],
     (err, rows, fields) => {
-      if (!err){
-         res.status(200).json({
-        success:true,
-        data: rows
-      })
-      }
-      else{
+      if (!err) {
         res.status(200).json({
-        success: false,
-        data: { message: "Something went to wrong!" + err },
-      });
+          success: true,
+          data: rows,
+        });
+      } else {
+        res.status(200).json({
+          success: false,
+          data: { message: "Something went to wrong!" + err },
+        });
       }
     }
   );
@@ -69,21 +67,26 @@ exports.DeleteById = (req, res) => {
 exports.CreateUser = (req, res) => {
   let tempData = req.body;
   tempData.password = md5(tempData.password);
+  var salry = tempData.salary?tempData.salary:'0';
+
   mysqlConnection.query(
     "INSERT INTO `users`(`name`, `email`, `password`) VALUES (?,?,?)",
     [tempData.name, tempData.email, tempData.password],
     function (err, rows) {
+      mysqlConnection.query(
+        "INSERT INTO `payroll`(`userId`, `salary`) VALUES (?,?)",
+        [rows.insertId,salry]
+      );
       if (!err) {
         res.status(200).json({
-          success:true,
-          data: rows
-        })
-      }
-      else{
+          success: true,
+          data: rows,
+        });
+      } else {
         res.status(200).json({
-        success: false,
-        data: { message: "Something went to wrong!" + err },
-      });
+          success: false,
+          data: { message: "Something went to wrong!" + err },
+        });
       }
     }
   );
@@ -92,7 +95,7 @@ exports.CreateUser = (req, res) => {
 //Update an User
 exports.UpdateUser = (req, res) => {
   let tempData = req.body;
-  if(tempData.password) tempData.password = md5(tempData.password);
+  if (tempData.password) tempData.password = md5(tempData.password);
 
   mysqlConnection.query(
     "UPDATE `users` SET `name`=?,`email`=?,`password`=? where `id`=?",
@@ -100,11 +103,10 @@ exports.UpdateUser = (req, res) => {
     function (err, rows, fields) {
       if (!err) {
         res.status(200).json({
-          success:true,
-          data: rows
-        })
-      }
-      else {
+          success: true,
+          data: rows,
+        });
+      } else {
         res.status(200).json({
           success: false,
           data: { message: "Something went to wrong!" + err },
